requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.56.0
fsspec==2025.2.0
idna==3.10
importlib-resources==6.4.5
jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
pillow==10.4.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyntcore==2023.4.3.0
pyparsing==3.1.4
pyrealsense2==2.55.1.6486
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
robotpy-cscore==2023.4.3.0
robotpy-wpinet==2023.4.3.0
robotpy-wpiutil==2023.4.3.0
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.78
ultralytics-thop==2.0.14
urllib3==2.2.3
zipp==3.20.2
